Anomaly Detection/requirements.txt
# Python dependencies for anomaly detection layer

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0
torch>=1.10.0  # Optional, for AutoEncoder detector

# Message queue
pyzmq>=22.0.0

# Configuration
PyYAML>=5.4.0

# Optional: for better performance
# pandas>=1.3.0

